geo_location/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
attrs==24.3.0
branca==0.8.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
folium==0.19.2
frozenlist==1.5.0
idna==3.10
ipinfo==5.1.1
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.0
propcache==0.2.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
xyzservices==2024.9.0
yarl==1.18.3
